+++ src/validateInvocation.ts
@@ -25,4 +25,5 @@ export default async function validateInvocation(
   const apiClient = createAPIClient(config);
 
   await apiClient.verifyAuthentication();
+  await apiClient.verifyPipelineAuthentication();
 }